Recipe data now resolves and redirects to the list on error.
When loading a recipe detail page, the recipe now resolves ahead of
time. If there is an error, it instead redirects to the recipe list
view. This means that errors caused by being unable to read a recipe
prevent loading.

This is a temporary fix; I need to figure out a better way of handling
it that uses the CanActivate guard somehow.