Fixed problem where FAB buttons hovered over content.
Essentially, when scrolled all the way down, some views would have FAB
buttons covering content. To fix this, I added margins on the bottoms of
the recipe-list and recipe-detail views so that the FAB would always
have empty, content-less space to hover over even when the view was
completely scrolled down.